Recompile commentaries with some fixes for trailing unknown characters
What was happening is that some characters were being replaced across
grapheme boundaries. From the Elixir docs:

> Be aware that this function can replace within or across grapheme boundaries. For example, take the grapheme "é" which is made of the characters "e" and the acute accent. The following will replace only the letter "e", moving the accent to the letter "o":

https://hexdocs.pm/elixir/1.16.2/String.html#replace/4

This resulted in unknown characters being printed in the middle of
lemmata. Even when we anchored the replacement regex to the end
of the string, sometimes unknown characters would result. So now
we also replace invalid characters (which *should* only appear
at the end of the string) with an empty string to clean everything up.